pyproject.toml
[project]
name = "flatsat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitpython>=3.1.45",
    "opencv-python",
    "numpy",
]